Deprecate the EffectRouterBuilder
